refactor name convention after merge with User branch
