+++ inc/Server.hpp
@@ -29,8 +29,8 @@
 # include <fcntl.h>
 # include <poll.h>
 # include <arpa/inet.h> // for inet_ntop()
-//# include "User.hpp"
-//# include "Channel.hpp"
+# include "User.hpp"
+# include "Channel.hpp"
 
 
 #define MIN_PORT_NUMBER	49152      //1025 Registered Ports (1.024 - 49.151) -----   Dynamically Allocated Ports (49.152 - 65.535):
